refactor: 로그인이 필요 없는 로직 위치 이동
user는 로그인이 필요한 컨트롤러라서 로그인이 필요없는 함수는 login controller로 위치 이동
